fix get_transformer function and amend imports
